refactoring Move from Chess.js to its own file; created types.js to avoid circular dependencies; some changes noted in chess.js.api.md
