+++ src/types.ts
@@ -1,0 +1,82 @@
+/**
+ * Shared types and constants used across chess.js modules
+ */
+
+export const WHITE = 'w'
+export const BLACK = 'b'
+
+export const PAWN = 'p'
+export const KNIGHT = 'n'
+export const BISHOP = 'b'
+export const ROOK = 'r'
+export const QUEEN = 'q'
+export const KING = 'k'
+
+export type Color = 'w' | 'b'
+export type PieceSymbol = 'p' | 'n' | 'b' | 'r' | 'q' | 'k'
+
+// prettier-ignore
+export type Square =
+    'a8' | 'b8' | 'c8' | 'd8' | 'e8' | 'f8' | 'g8' | 'h8' |
+    'a7' | 'b7' | 'c7' | 'd7' | 'e7' | 'f7' | 'g7' | 'h7' |
+    'a6' | 'b6' | 'c6' | 'd6' | 'e6' | 'f6' | 'g6' | 'h6' |
+    'a5' | 'b5' | 'c5' | 'd5' | 'e5' | 'f5' | 'g5' | 'h5' |
+    'a4' | 'b4' | 'c4' | 'd4' | 'e4' | 'f4' | 'g4' | 'h4' |
+    'a3' | 'b3' | 'c3' | 'd3' | 'e3' | 'f3' | 'g3' | 'h3' |
+    'a2' | 'b2' | 'c2' | 'd2' | 'e2' | 'f2' | 'g2' | 'h2' |
+    'a1' | 'b1' | 'c1' | 'd1' | 'e1' | 'f1' | 'g1' | 'h1'
+
+export type Piece = {
+  color: Color
+  type: PieceSymbol
+}
+
+export type InternalMove = {
+  color: Color
+  from: number
+  to: number
+  piece: PieceSymbol
+  captured?: PieceSymbol
+  promotion?: PieceSymbol
+  flags: number
+}
+
+export const FLAGS: Record<string, string> = {
+  NORMAL: 'n',
+  CAPTURE: 'c',
+  BIG_PAWN: 'b',
+  EP_CAPTURE: 'e',
+  PROMOTION: 'p',
+  KSIDE_CASTLE: 'k',
+  QSIDE_CASTLE: 'q',
+  NULL_MOVE: '-',
+}
+
+export const BITS: Record<string, number> = {
+  NORMAL: 1,
+  CAPTURE: 2,
+  BIG_PAWN: 4,
+  EP_CAPTURE: 8,
+  PROMOTION: 16,
+  KSIDE_CASTLE: 32,
+  QSIDE_CASTLE: 64,
+  NULL_MOVE: 128,
+}
+
+// Extracts the zero-based rank of an 0x88 square.
+export function rank(square: number): number {
+  return square >> 4
+}
+
+// Extracts the zero-based file of an 0x88 square.
+export function file(square: number): number {
+  return square & 0xf
+}
+
+// Converts a 0x88 square to algebraic notation.
+export function algebraic(square: number): Square {
+  const f = file(square)
+  const r = rank(square)
+  return ('abcdefgh'.substring(f, f + 1) +
+    '87654321'.substring(r, r + 1)) as Square
+}
